OPERATOR-458 Move StorageCluster creation code to different file
